Alles double

+++ src/BaseCalculation.java
@@ -1,10 +1,10 @@
 public abstract class BaseCalculation {
     protected int elements;
-    protected float lambda;
-    public BaseCalculation(int elements, float lambda){
+    protected double lambda;
+    public BaseCalculation(int elements, double lambda){
         this.elements = elements;
         this.lambda = lambda;
     };
-    abstract float calculateParallelMTTF();
-    abstract float calculateSerialMTTF();
+    abstract double calculateParallelMTTF();
+    abstract double calculateSerialMTTF();
 }
